+++ srcs/socket/Send.cpp
@@ -6,7 +6,7 @@
 /*   By: fmonbeig <fmonbeig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/16 11:58:15 by fmonbeig          #+#    #+#             */
-/*   Updated: 2022/03/16 14:17:12 by fmonbeig         ###   ########.fr       */
+/*   Updated: 2022/03/17 15:27:58 by fmonbeig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,10 @@ static int	contentSize(char *content)
 void	sendMessage(int link, char *buff)
 {
 	//TODO we have to analyse the message from the client and create the right content
-	char	content[] = "HTTP/1.1 200 OK\nContent-Type: text/plain\nContent-Length: 12\n\nHello world!";
+	// char	content[] = "HTTP/1.1 200 OK\nContent-Type: text/plain\nContent-Length: 12\n\nHello world!";
+	char content[] = "HTTP/1.1 200 OK\nContent-Type: text/html\nContent-Length: 202\n\n"
+		"<form method=\"post\" enctype=\"multipart/form-data\"><div><label for=\"file\">select file</label><input type=\"file\" id=\"file\" name=\"file\" multiple></div><div><button>Envoyer</button></div></form>";
+
 	int		ret;
 	int		temp;
